:construction: Continuation of the implementation of commands and reorganization of the code and files.

+++ src/Entity/Carrier.php
@@ -22,7 +22,7 @@ class Carrier implements Stringable
     private ?string $description = null;
 
     #[ORM\Column]
-    private ?float $price = null;
+    private ?int $price = null;
 
 public function __toString()
 {
@@ -60,12 +60,12 @@ public function __toString()
         return $this;
     }
 
-    public function getPrice(): ?float
+    public function getPrice(): ?int
     {
         return $this->price;
     }
 
-    public function setPrice(float $price): static
+    public function setPrice(int $price): static
     {
         $this->price = $price;
